+++ lets-fiware.sh
@@ -28,7 +28,7 @@
 
 set -Ceuo pipefail
 
-VERSION=0.12.0
+VERSION=0.12.0-next
 
 #
 # Syslog info