+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## FIWARE Big Bang v0.11.0-next
 
+-   Update documentaion (#157)
 -   Update Quantumleap to 0.8.3 (#158)
 
 ## FIWARE Big Bang v0.11.0 - 10 Feb, 2022