+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## FIWARE Big Bang v0.12.0-next
+
 ## FIWARE Big Bang v0.12.0 - 13 Feb, 2022
 
 -   Support Apache Zeppelin (#160)